Added support for procedural dome generation. Started implementing generation of averaged terrain normals. Clean up a few warnings.

+++ inc/vapor/input/Mouse.h
@@ -147,9 +147,9 @@ struct MouseButtonEvent : public MouseEvent
 
 struct MouseWheelEvent : public MouseEvent
 {
-	MouseWheelEvent( int delta );
+	MouseWheelEvent( short delta );
 
-	int delta;
+	short delta;
 };
 
 //-----------------------------------//
